Add suggestion changes
